dev: fixed naming error
