+++ src/components/SingInForm/SignInForm.tsx
@@ -7,13 +7,14 @@ import BaseButton from 'uikit/BaseButton'
 import { EBaseButtonMode } from 'uikit/BaseButton/BaseButton'
 import BaseInput from 'uikit/BaseInput'
 
-import styles from './SingUpForm.module.css'
-export interface SingUpFormProps {
+import styles from './SignInForm.module.css'
+
+export interface SignInFormProps {
   userFunction: () => void
-  onSignUp: (data: ISignUpData) => void
+  onSignIn: (data: ISignUpData) => void
 }
 
-const SingUpForm: FC<SingUpFormProps> = ({ userFunction, onSignUp }) => {
+const SignInForm: FC<SignInFormProps> = ({ userFunction, onSignIn }) => {
   const { isInProgress } = useAppSelector(state => state.auth)
 
   const email = useInput('', {
@@ -26,36 +27,19 @@ const SingUpForm: FC<SingUpFormProps> = ({ userFunction, onSignUp }) => {
     minLength: { value: 8, message: 'Minimum password length 8 symbols' },
   })
 
-  const confirmPassword = useInput('', {
-    isEmpty: {
-      value: true,
-      message: 'Please confirm your password',
-    },
-    isValueMatched: {
-      value: true,
-      comparisonValue: password.value,
-      message: 'Password mismatch',
-    },
-  })
-
   const submitHandler = () => {
     email.onBlur()
     password.onBlur()
-    confirmPassword.onBlur()
 
-    if (
-      !email.errors.length &&
-      !password.errors.length &&
-      !confirmPassword.errors.length
-    ) {
-      onSignUp({ email: String(email.value), password: String(password.value) })
+    if (!email.errors.length && !password.errors.length) {
+      onSignIn({ email: String(email.value), password: String(password.value) })
     }
   }
 
   return (
     <div className={styles.root}>
       <span className={styles.title}>
-        Sign <span className={styles['red-text']}>up</span>
+        Sign <span className={styles['red-text']}>in</span>
       </span>
 
       <BaseInput
@@ -84,23 +68,10 @@ const SingUpForm: FC<SingUpFormProps> = ({ userFunction, onSignUp }) => {
         isDirty={password.isDirty}
       />
 
-      <BaseInput
-        dataTestId={'confirm-password-input'}
-        type={'password'}
-        name={'confirm-password'}
-        placeholder={'password confirm'}
-        value={confirmPassword.value}
-        onChange={confirmPassword.onChange}
-        onBlur={confirmPassword.onBlur}
-        errors={confirmPassword.errors}
-        errorWhite
-        isDirty={confirmPassword.isDirty}
-      />
-
       <div className={styles.buttons}>
         <BaseButton
           userStyles={styles.button}
-          label={'Sign up'}
+          label={'Sign in'}
           mode={EBaseButtonMode.CONTAINED_RED}
           onClick={submitHandler}
           loading={isInProgress}
@@ -117,4 +88,4 @@ const SingUpForm: FC<SingUpFormProps> = ({ userFunction, onSignUp }) => {
   )
 }
 
-export default SingUpForm
+export default SignInForm
